+++ lib/identity_document_detection.dart
@@ -1,3 +1,5 @@
 library identity_document_detection;
 
-export 'src/view/identity_document_detection.dart';
+export 'src/view/identity_detector.dart';
+export 'src/controller/detector_service.dart';
+export 'src/model/options.dart';