+++ lib/src/controller/detector_service.dart
@@ -7,12 +7,15 @@ import 'dart:io';
 import 'dart:isolate';
 
 import 'package:camera/camera.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:identity_document_detection/src/controller/utils/image_utils.dart';
-import 'package:identity_document_detection/src/model/recognition.dart';
-import 'package:image/image.dart' as image_lib;
+import 'package:identity_document_detection/src/model/identity_document.dart';
+import 'package:identity_document_detection/src/model/options.dart';
+import 'package:image/image.dart' as img;
 import 'package:tflite_flutter/tflite_flutter.dart';
+
+part 'package:identity_document_detection/src/model/command.dart';
+part 'package:identity_document_detection/src/controller/detector_server.dart';
 
 ///////////////////////////////////////////////////////////////////////////////
 // **WARNING:** This is not production code and is only intended to be used for
@@ -44,39 +47,22 @@ import 'package:tflite_flutter/tflite_flutter.dart';
 //
 ///////////////////////////////////////////////////////////////////////////////
 
-/// All the command codes that can be sent and received between [Detector] and
-/// [_DetectorServer].
-enum _Codes {
-  init,
-  busy,
-  ready,
-  detect,
-  result,
-}
-
-/// A command sent between [Detector] and [_DetectorServer].
-class _Command {
-  const _Command(this.code, {this.args});
-
-  final _Codes code;
-  final List<Object>? args;
-}
-
 /// A Simple Detector that handles object detection via Service
 ///
 /// All the heavy operations like pre-processing, detection, ets,
 /// are executed in a background isolate.
 /// This class just sends and receives messages to the isolate.
-class Detector {
+///
+class IDController {
   static const String _modelPath = 'packages/identity_document_detection/assets/model/document.tflite';
-  // static const String _modelPath = 'assets/model/document.tflite';
   static const String _labelPath = 'packages/identity_document_detection/assets/model/classes.txt';
 
-  Detector._(this._isolate, this._interpreter, this._labels);
+  IDController._(this._isolate, this._interpreter, this._labels, this._options);
 
   final Isolate _isolate;
   late final Interpreter _interpreter;
   late final List<String> _labels;
+  final IDOptions _options;
 
   // To be used by detector (from UI) to send message to our Service ReceivePort
   late final SendPort _sendPort;
@@ -85,19 +71,17 @@ class Detector {
 
   // // Similarly, StreamControllers are stored in a queue so they can be handled
   // // asynchronously and serially.
-  final StreamController<Map<String, dynamic>> resultsStream = StreamController<Map<String, dynamic>>();
+  final StreamController<List<IdentityDocument>> _stream = StreamController<List<IdentityDocument>>();
+  Stream<List<IdentityDocument>> get stream => _stream.stream;
 
   /// Open the database at [path] and launch the server on a background isolate..
-  static Future<Detector> start() async {
+  static Future<IDController> initialize([IDOptions confidence = IDOptions.byDefault]) async {
     final ReceivePort receivePort = ReceivePort();
     // sendPort - To be used by service Isolate to send message to our ReceiverPort
     final Isolate isolate = await Isolate.spawn(_DetectorServer._run, receivePort.sendPort);
-
-    final Detector result = Detector._(
-      isolate,
-      await _loadModel(),
-      await _loadLabels(),
-    );
+    final model = await _loadModel();
+    final labels = await _loadLabels();
+    final IDController result = IDController._(isolate, model, labels, confidence);
     receivePort.listen((message) {
       result._handleCommand(message as _Command);
     });
@@ -119,7 +103,9 @@ class Detector {
   }
 
   static Future<List<String>> _loadLabels() async {
-    return (await rootBundle.loadString(_labelPath)).split('\n');
+    final fileContent = await rootBundle.loadString(_labelPath);
+    final labels = fileContent.split('\n');
+    return labels;
   }
 
   /// Starts CameraImage processing
@@ -149,203 +135,29 @@ class Detector {
         ]));
       case _Codes.ready:
         _isReady = true;
+        break;
       case _Codes.busy:
         _isReady = false;
+        break;
       case _Codes.result:
         _isReady = true;
-        resultsStream.add(command.args?[0] as Map<String, dynamic>);
+        var docs = command.args as List<IdentityDocument>;
+        if (_options.detection == IDTypeDetection.multiple) {
+          docs = docs.where((recognition) => recognition.score > _options.confidence).toList();
+        } else {
+          docs = [docs.reduce((A, B) => A.score > B.score && B.score > _options.confidence ? A : B)];
+          if (docs.first.score < _options.confidence) docs = [];
+        }
+        _options.onDocumentDetect?.call(docs);
+        _stream.add(docs);
+        break;
       default:
-        debugPrint('Detector unrecognized command: ${command.code}');
+        break;
     }
   }
 
   /// Kills the background isolate and its detector server.
   void stop() {
     _isolate.kill();
-  }
-}
-
-/// The portion of the [Detector] that runs on the background isolate.
-///
-/// This is where we use the new feature Background Isolate Channels, which
-/// allows us to use plugins from background isolates.
-class _DetectorServer {
-  /// Input size of image (height = width = 320)
-  static const int mlModelInputSize = 320;
-
-  /// Result confidence threshold
-  static const double confidence = 0.4;
-  Interpreter? _interpreter;
-  List<String>? _labels;
-
-  _DetectorServer(this._sendPort);
-
-  final SendPort _sendPort;
-
-  // ----------------------------------------------------------------------
-  // Here the plugin is used from the background isolate.
-  // ----------------------------------------------------------------------
-
-  /// The main entrypoint for the background isolate sent to [Isolate.spawn].
-  static void _run(SendPort sendPort) {
-    ReceivePort receivePort = ReceivePort();
-    final _DetectorServer server = _DetectorServer(sendPort);
-    receivePort.listen((message) async {
-      final _Command command = message as _Command;
-      await server._handleCommand(command);
-    });
-    // receivePort.sendPort - used by UI isolate to send commands to the service receiverPort
-    sendPort.send(_Command(_Codes.init, args: [receivePort.sendPort]));
-  }
-
-  /// Handle the [command] received from the [ReceivePort].
-  Future<void> _handleCommand(_Command command) async {
-    switch (command.code) {
-      case _Codes.init:
-        // ----------------------------------------------------------------------
-        // The [RootIsolateToken] is required for
-        // [BackgroundIsolateBinaryMessenger.ensureInitialized] and must be
-        // obtained on the root isolate and passed into the background isolate via
-        // a [SendPort].
-        // ----------------------------------------------------------------------
-        RootIsolateToken rootIsolateToken = command.args?[0] as RootIsolateToken;
-        // ----------------------------------------------------------------------
-        // [BackgroundIsolateBinaryMessenger.ensureInitialized] for each
-        // background isolate that will use plugins. This sets up the
-        // [BinaryMessenger] that the Platform Channels will communicate with on
-        // the background isolate.
-        // ----------------------------------------------------------------------
-        BackgroundIsolateBinaryMessenger.ensureInitialized(rootIsolateToken);
-        _interpreter = Interpreter.fromAddress(command.args?[1] as int);
-        _labels = command.args?[2] as List<String>;
-        _sendPort.send(const _Command(_Codes.ready));
-      case _Codes.detect:
-        _sendPort.send(const _Command(_Codes.busy));
-        _convertCameraImage(command.args?[0] as CameraImage);
-      default:
-        debugPrint('_DetectorService unrecognized command ${command.code}');
-    }
-  }
-
-  void _convertCameraImage(CameraImage cameraImage) {
-    var preConversionTime = DateTime.now().millisecondsSinceEpoch;
-
-    convertCameraImageToImage(cameraImage).then((image) {
-      if (image != null) {
-        if (Platform.isAndroid) {
-          image = image_lib.copyRotate(image, angle: 90);
-        }
-
-        final results = analyseImage(image, preConversionTime);
-        _sendPort.send(_Command(_Codes.result, args: [results]));
-      }
-    });
-  }
-
-  Map<String, dynamic> analyseImage(image_lib.Image? image, int preConversionTime) {
-    var conversionElapsedTime = DateTime.now().millisecondsSinceEpoch - preConversionTime;
-
-    var preProcessStart = DateTime.now().millisecondsSinceEpoch;
-
-    /// Pre-process the image
-    /// Resizing image for model [300, 300]
-    final imageInput = image_lib.copyResize(
-      image!,
-      width: mlModelInputSize,
-      height: mlModelInputSize,
-    );
-
-    // Creating matrix representation, [300, 300, 3]
-    final imageMatrix = List.generate(
-      imageInput.height,
-      (y) => List.generate(
-        imageInput.width,
-        (x) {
-          final pixel = imageInput.getPixel(x, y);
-          return [pixel.r, pixel.g, pixel.b];
-        },
-      ),
-    );
-    var preProcessElapsedTime = DateTime.now().millisecondsSinceEpoch - preProcessStart;
-
-    var inferenceTimeStart = DateTime.now().millisecondsSinceEpoch;
-
-    final output = _runInference(imageMatrix);
-    // Scores
-    final scores = output.first.first as List<double>;
-    print('scores: $scores');
-    // Location
-    final locationsRaw = output.elementAt(1).first as List<List<double>>;
-    final List<Rect> locations = locationsRaw
-        .map((list) => list.map((value) => (value * mlModelInputSize)).toList())
-        .map((rect) => Rect.fromLTRB(rect[1], rect[0], rect[3], rect[2]))
-        .toList();
-    print('locations: $locations');
-    // Classes
-    final classesRaw = output.last.first as List<double>;
-    final classes = classesRaw.map((value) => value.toInt()).toList();
-
-    print('classes: $classes');
-
-    // Number of detections
-    final numberOfDetectionsRaw = output[2].first as double;
-    final numberOfDetections = numberOfDetectionsRaw.toInt();
-    print('numberOfDetections: $numberOfDetections');
-
-    final List<String> classification = [];
-    for (var i = 0; i < numberOfDetections; i++) {
-      classification.add(_labels![classes[i]]);
-    }
-
-    /// Generate recognitions
-    List<Recognition> recognitions = [];
-    for (int i = 0; i < numberOfDetections; i++) {
-      // Prediction score
-      var score = scores[i];
-      // Label string
-      var label = classification[i];
-
-      if (score > confidence) {
-        recognitions.add(
-          Recognition(i, label, score, locations[i]),
-        );
-      }
-    }
-
-    var inferenceElapsedTime = DateTime.now().millisecondsSinceEpoch - inferenceTimeStart;
-
-    var totalElapsedTime = DateTime.now().millisecondsSinceEpoch - preConversionTime;
-
-    return {
-      "recognitions": recognitions,
-      "stats": <String, String>{
-        'Conversion time:': conversionElapsedTime.toString(),
-        'Pre-processing time:': preProcessElapsedTime.toString(),
-        'Inference time:': inferenceElapsedTime.toString(),
-        'Total prediction time:': totalElapsedTime.toString(),
-        'Frame': '${image.width} X ${image.height}',
-      },
-    };
-  }
-
-  /// Object detection main function
-  List<List<Object>> _runInference(List<List<List<num>>> imageMatrix) {
-    // Set input tensor [1, 320, 320, 3]
-    final input = [imageMatrix];
-
-    // Set output tensor
-    // Locations: [1, 10, 4]
-    // Classes: [1, 10],
-    // Scores: [1, 10],
-    // Number of detections: [1]
-    final output = {
-      0: [List<double>.filled(25, 0.0)],
-      1: [List<List<double>>.filled(25, List<double>.filled(4, 0.0))],
-      2: [0.0],
-      3: [List<double>.filled(25, 0.0)],
-    };
-
-    _interpreter!.runForMultipleInputs([input], output);
-    return output.values.toList();
   }
 }